Move padlock.h to library
`padlock.h` is actually an internal header and is moved accordingly.
